feat: AdmProductImageControllerTest
